file download retry until resource is consistent
changes similar to upload to retry while we are waiting for the file to be consistent
moved some code around since it is used in generic places now
