+++ ddsc/core/download.py
@@ -1,7 +1,8 @@
 import os
-from ddsc.core.util import ProgressPrinter
+from ddsc.core.util import ProgressPrinter, ProjectStatusMonitor
 from ddsc.core.filedownloader import FileDownloader
 from ddsc.core.pathfilter import PathFilteredProject
+from ddsc.core.ddsapi import retry_until_resource_is_consistent
 
 
 class ProjectDownload(object):
@@ -23,6 +24,7 @@ class ProjectDownload(object):
         self.path_filter = path_filter
         self.file_download_pre_processor = file_download_pre_processor
         self.watcher = None
+        self.project_status_monitor = None
 
     def run(self):
         """
@@ -41,6 +43,7 @@ class ProjectDownload(object):
         path_filtered_project.run(project)  # calls visit_project, visit_folder, visit_file in RemoteContentCounter
 
         self.watcher = ProgressPrinter(counter.count, msg_verb='downloading')
+        self.project_status_monitor = ProjectStatusMonitor(self.watcher, action_name='downloading')
         path_filtered_project = PathFilteredProject(self.path_filter, self)
         path_filtered_project.run(project)  # calls visit_project, visit_folder, visit_file below
         self.watcher.finished()
@@ -84,7 +87,8 @@ class ProjectDownload(object):
         if self.file_download_pre_processor:
             self.file_download_pre_processor.run(self.remote_store.data_service, item)
         path = os.path.join(self.dest_directory, item.remote_path)
-        url_json = self.remote_store.data_service.get_file_url(item.id).json()
+        get_file_url = GetFileUrl(self.remote_store.data_service, item)
+        url_json = retry_until_resource_is_consistent(get_file_url.run, self.project_status_monitor)
         downloader = FileDownloader(self.remote_store.config, item, url_json, path, self.watcher)
         downloader.run()
         ProjectDownload.check_file_size(item, path)
@@ -137,3 +141,12 @@ class RemoteContentCounter(object):
         :return:
         """
         self.count += item.size
+
+
+class GetFileUrl(object):
+    def __init__(self, data_service, item):
+        self.data_service = data_service
+        self.item = item
+
+    def run(self):
+        return self.data_service.get_file_url(self.item.id).json()
