+++ ddsc/core/tests/test_fileuploader.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 from ddsc.core.fileuploader import ParallelChunkProcessor, upload_async, FileUploadOperations, \
-    RESOURCE_NOT_CONSISTENT_RETRY_SECONDS, ProjectStatusMonitor
+    RESOURCE_NOT_CONSISTENT_RETRY_SECONDS
 from ddsc.core.ddsapi import DSResourceNotConsistentError, DataServiceError
 import requests
 from mock import MagicMock, Mock, patch, call
@@ -200,31 +200,3 @@ class TestFileUploadOperations(TestCase):
         fop = FileUploadOperations(data_service, MagicMock())
         with self.assertRaises(DataServiceError):
             fop.create_upload(project_id='12', path_data=path_data, hash_data=MagicMock())
-
-
-class TestProjectStatusMonitor(TestCase):
-    def test_started_waiting_debounces(self):
-        watcher = MagicMock()
-        monitor = ProjectStatusMonitor(watcher)
-        monitor.started_waiting()
-        self.assertEqual(1, watcher.start_waiting.call_count)
-        watcher.start_waiting.assert_has_calls([
-            call('Waiting for project to become ready for uploading')
-        ])
-        monitor.started_waiting()
-        monitor.started_waiting()
-        self.assertEqual(1, watcher.start_waiting.call_count)
-        watcher.start_waiting.assert_has_calls([
-            call('Waiting for project to become ready for uploading')
-        ])
-
-    def test_done_waiting(self):
-        watcher = MagicMock()
-        monitor = ProjectStatusMonitor(watcher)
-        monitor.started_waiting()
-        monitor.done_waiting()
-        self.assertEqual(1, watcher.start_waiting.call_count)
-        watcher.start_waiting.assert_has_calls([
-            call('Waiting for project to become ready for uploading'),
-        ])
-        self.assertEqual(1, watcher.done_waiting.call_count)
